requirements.txt
Flask
PyPDF2
google-generativeai
streamlit